requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-multipart==0.0.9
Jinja2==3.1.3
google-generativeai==0.3.2
pandas==2.2.1
openai==1.12.0
openpyxl==3.1.2
python-dotenv==1.0.1